+++ cmd/stratumd/src/stratum_rpc.rs
